Commit de tipos de datos en los modelos

+++ Danchi/Models/Administrador.cs
@@ -4,7 +4,7 @@
     {
         public int IdAdministrador { get; set; }
         public string Nombre { get; set; }
-        public int CelularAdministrador { get; set; }
+        public decimal CelularAdministrador { get; set; }
         public string CorreoElectronicoAd { get; set; }
     }
 }
