+++ Danchi/Models/Residente.cs
@@ -5,7 +5,7 @@
         public int IdResidente { get; set; }
         public string NumApartamento { get; set; }
         public string CorreoElectronico { get; set; }
-        public int CelularResidente { get; set; }
+        public decimal CelularResidente { get; set; }
         public string Nombre { get; set; }
     }
 }
